+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.64 (2024-9-26)
+
+- fix: set default order index for categories and deployments
+- fix: project settings errors when saving project
+
 ## 0.4.63 (2024-9-26)
 
 - feat: support sort category and deployments